requirements.txt
beautifulsoup4==4.12.3
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.1
Requests==2.31.0
reflex==0.4.0
